+++ config/initializers/ckeditor.rb
@@ -1,0 +1,8 @@
+Ckeditor.setup do |config|
+  require "ckeditor/orm/active_record"
+
+  config.default_per_page = 24
+
+  config.assets_languages = ['en']
+
+end