+++ lib/community_engine/admin/pages.rb
@@ -1,0 +1,3 @@
+ActiveAdmin.register Page do
+  # everything happens here :D
+end