+++ lib/community_engine/admin/tags.rb
@@ -1,0 +1,3 @@
+ActiveAdmin.register ActsAsTaggableOn::Tag, as: "Tag" do
+  # everything happens here :D
+end