+++ lib/community_engine/admin/users.rb
@@ -1,0 +1,3 @@
+ActiveAdmin.register User do
+  # everything happens here :D
+end